+++ src/meltano/core/plugin/airflow.py
@@ -6,6 +6,7 @@ import time
 import os
 from . import PluginInstall, PluginType
 
+from meltano.core.error import SubprocessError
 from meltano.core.behavior.hookable import hook
 from meltano.core.plugin.config_service import PluginConfigService
 from meltano.core.plugin.settings_service import PluginSettingsService